fix counting

+++ SHAChecker/models.py
@@ -4,7 +4,7 @@ from django.db import models
 # Create your models here.
 class CountSHA(models.Model):
     key = models.CharField(max_length=255, unique=True)
-    value = models.IntegerField(unique=False)
+    value = models.FloatField(unique=False)
 
     def __str__(self):
         return "SHA256: " + self.key + "     Value: " + str(self.value)
